requirements.txt
docopt==0.6.2
et-xmlfile==1.1.0
openpyxl==3.1.2
pyelftools==0.29
